move profile folder from js_modules
