+++ src/common/store/actions/subscriptionAction.js
@@ -2,7 +2,7 @@ import { useSelector, useDispatch } from 'react-redux';
 import { useToast } from '@chakra-ui/react';
 import { CANCEL_SUBSCRIPTION, FETCH_SUBSCRIPTIONS, IS_LOADING } from '../types';
 import bc from '../../services/breathecode';
-import profileHandlers from '../../../js_modules/profile/Subscriptions/handlers';
+import profileHandlers from '../../components/Profile/Subscriptions/handlers';
 
 const useSubscriptionsHandler = () => {
   const state = useSelector((st) => st.subscriptionsReducer);
@@ -10,7 +10,7 @@ const useSubscriptionsHandler = () => {
   const dispatch = useDispatch();
   const {
     getPlanOffer,
-  } = profileHandlers({});
+  } = profileHandlers();
 
   const fetchSubscriptions = () => new Promise((resolve, reject) => {
     dispatch({
